requirements.txt
aiodns==3.0.0
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-cache==1.1.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
black==22.12.0
Brotli==1.0.9
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.3.0
fake-headers==1.0.2
fastapi==0.89.1
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
httptools==0.5.0
idna==3.4
limits==2.8.0
lxml==4.9.2
multidict==6.0.4
mutagen==1.46.0
mypy-extensions==0.4.3
orjson==3.8.5
packaging==22.0
pathspec==0.10.3
platformdirs==2.6.2
pycares==4.3.0
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.4
python-dotenv==0.21.0
PyYAML==6.0
six==1.16.0
slowapi==0.1.7
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
webencodings==0.5.1
websockets==10.4
wrapt==1.14.1
yarl==1.8.2
yt-dlp==2023.1.6
